Improve error printing in jline repl
